# Disable t1ha indirect functions to fix x86_64-unknown-linux-musl 