--- a/fasthash/benches/hash.rs
+++ b/fasthash/benches/hash.rs
@@ -125,15 +125,6 @@ fn bench_hash64(c: &mut Criterion) {
     .with_function("spooky::hash64", move |b, &&size| {
         b.iter(|| spooky::hash64_with_seed(&DATA[..size], SEED));
     })
-    .with_function("t1ha0::hash64", move |b, &&size| {
-        b.iter(|| t1ha0::Hash64::hash_with_seed(&DATA[..size], SEED));
-    })
-    .with_function("t1ha1::hash64", move |b, &&size| {
-        b.iter(|| t1ha1::Hash64::hash_with_seed(&DATA[..size], SEED));
-    })
-    .with_function("t1ha2::hash64_atonce", move |b, &&size| {
-        b.iter(|| t1ha2::Hash64AtOnce::hash_with_seed(&DATA[..size], SEED));
-    })
     .with_function("xx::hash64", move |b, &&size| {
         b.iter(|| xx::hash64_with_seed(&DATA[..size], SEED));
     })
@@ -143,6 +134,18 @@ fn bench_hash64(c: &mut Criterion) {
     .with_function("highway::hash64", move |b, &&size| {
         b.iter(|| highway::hash64_with_seed(&DATA[..size], [SEED, SEED, SEED, SEED]));
     });
+
+    #[cfg(feature = "t1ha")] {
+        bench = bench.with_function("t1ha0::hash64", move |b, &&size| {
+            b.iter(|| t1ha0::Hash64::hash_with_seed(&DATA[..size], SEED));
+        })
+            .with_function("t1ha1::hash64", move |b, &&size| {
+                b.iter(|| t1ha1::Hash64::hash_with_seed(&DATA[..size], SEED));
+            })
+            .with_function("t1ha2::hash64_atonce", move |b, &&size| {
+                b.iter(|| t1ha2::Hash64AtOnce::hash_with_seed(&DATA[..size], SEED));
+            });
+    }
 
     if cfg!(any(feature = "sse4.2", target_feature = "sse4.2")) {
         bench = bench
@@ -189,15 +192,19 @@ fn bench_hash128(c: &mut Criterion) {
     .with_function("spooky::hash128", move |b, &&size| {
         b.iter(|| spooky::hash128_with_seed(&DATA[..size], SEED as u128));
     })
-    .with_function("t1ha2::hash128_atonce", move |b, &&size| {
-        b.iter(|| t1ha2::Hash128AtOnce::hash_with_seed(&DATA[..size], SEED));
-    })
     .with_function("xxh3::hash128", move |b, &&size| {
         b.iter(|| xxh3::hash128_with_seed(&DATA[..size], SEED));
     })
     .with_function("highway::hash128", move |b, &&size| {
         b.iter(|| highway::hash128_with_seed(&DATA[..size], [SEED, SEED, SEED, SEED]));
     });
+
+    #[cfg(feature = "t1ha")] {
+        bench = bench
+            .with_function("t1ha2::hash128_atonce", move |b, &&size| {
+                b.iter(|| t1ha2::Hash128AtOnce::hash_with_seed(&DATA[..size], SEED));
+            });
+    }
 
     if cfg!(any(feature = "sse4.2", target_feature = "sse4.2")) {
         bench = bench