--- a/fasthash/src/lib.rs
+++ b/fasthash/src/lib.rs
@@ -79,9 +79,10 @@ pub mod mum;
 pub mod murmur;
 pub mod murmur2;
 pub mod murmur3;
+#[cfg(feature = "t1ha")]
+pub mod t1ha;
 pub mod sea;
 pub mod spooky;
-pub mod t1ha;
 pub mod xx;
 pub mod xxh3;
 
@@ -97,8 +98,12 @@ pub use crate::murmur3::Hasher32 as Murmur3Hasher;
 #[doc(no_inline)]
 pub use crate::sea::Hasher64 as SeaHasher;
 pub use crate::spooky::{Hasher128 as SpookyHasherExt, Hasher64 as SpookyHasher};
-pub use crate::t1ha::{t1ha0, t1ha1, t1ha2};
-pub use crate::t1ha2::{Hasher128 as T1haHasherExt, Hasher128 as T1haHasher};
+cfg_if! {
+    if #[cfg(feature = "t1ha")] {
+        pub use crate::t1ha::{t1ha0, t1ha1, t1ha2};
+        pub use crate::t1ha2::{Hasher128 as T1haHasherExt, Hasher128 as T1haHasher};
+    }
+}
 pub use crate::xx::Hasher64 as XXHasher;
 cfg_if! {
     if #[cfg(target_pointer_width = "64")] {